Move app link form to own component
