+++ package.json
@@ -10,6 +10,7 @@
     "@segment/utm-params": "^2.0.0",
     "@types/enzyme": "^3.1.13",
     "@types/jest": "^23.3.1",
+    "@types/js-cookie": "^2.2.0",
     "@types/prop-types": "^15.5.5",
     "@types/ramda": "^0.25.42",
     "@types/react": "^16.4.18",
